Rename `configFiles.ts` to `studyFile.ts`
