+++ src/helpers/apiManager.ts
@@ -7,8 +7,8 @@ import { AnswerEntity } from "../entities/AnswerEntity";
 import { PingEntity } from "../entities/PingEntity";
 import { getAnswersAsync } from "./answers";
 import { User, storeUserAsync, getUserAsync } from "./asyncStorage/user";
-import { getStudyInfoAsync } from "./configFiles";
 import { getPingsAsync } from "./pings";
+import { getStudyInfoAsync } from "./studyFile";
 
 export async function getServerUrlAsync(): Promise<string> {
   return (await getStudyInfoAsync()).serverURL;
